Add files via upload
Entry data:
The name of the input file from which it is read is received as a program argument. The input files will have the following structure:
•	On the first line there will be a natural number N, representing the total number of shapes to be drawn in the drawing.
•	The following N lines will contain information about each of N forms: the position of a shape, its dimensions, and its color.
The colors you'll need to use will be rendered in the format: "#RGB A", where R, G, and B represent the red, green and blue pixels in hex (256 possible values: 00, 01, 02, ... , 09, 0A, 0B, ..., 0F, 10, 11, ..., FE, FF), and A is the decimal value of Alpha (opacity), taking values from 0 to 100.
Examples of colors that can appear in the input:
•	#110859 100
•	#1F0ab9 80
•	#0008Af 100

Example input:
An example of an input file along with the image generated by it can be found in the Github repository's media folder of theme 2 [7].

Output Data:
The output of the program should be a PNG image named "drawing.png" to present the forms described in the input.